GetChallengeByName endpoint and frontend admin can now delete challenges
