+++ frontend/src/AdminCategories.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+function AdminCategories() {
+  return <div>Admin Categories Management</div>;
+}
+export default AdminCategories;
